refactor: move OAuth grant factories from core to packages/oauth
Move createSessionGrant, createAuthorizationGrant, createRefreshTokenGrant
and their tests from core to packages/oauth. Core no longer exports any
OAuth-specific implementations — only protocol-agnostic abstractions.
