+++ packages/oauth/src/__tests__/refreshToken.test.mts
@@ -17,9 +17,12 @@ import { createSecretKey } from "node:crypto";
 import { SignJWT } from "jose";
 import { describe, expect, it } from "vitest";
 
-import { createSymmetricKeyStore } from "../../keys/KeyStore.mjs";
-import { createRefreshTokenGrant } from "../refreshToken.mjs";
-import type { GrantContext, GrantDependencies } from "../types.mjs";
+import {
+	createSymmetricKeyStore,
+	type GrantContext,
+	type GrantDependencies,
+} from "@o3co/auth-provider-core";
+import { createRefreshTokenGrant } from "../grants/refreshToken.mjs";
 
 const SECRET = "test-secret-at-least-32-chars!!";
 const keyStore = createSymmetricKeyStore(SECRET);
