+++ packages/oauth/src/grants/session.mts
@@ -13,15 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import type { ClientRepository } from "#/repositories/ClientRepository.mjs";
-
-import { generateToken, generateTokenResponse } from "./token.mjs";
-import type {
-	GrantContext,
-	GrantDependencies,
-	GrantHandler,
-	GrantHandlerResult,
-} from "./types.mjs";
+import {
+	generateToken,
+	generateTokenResponse,
+	type ClientRepository,
+	type GrantContext,
+	type GrantDependencies,
+	type GrantHandler,
+	type GrantHandlerResult,
+} from "@o3co/auth-provider-core";
 
 export const createSessionGrant = (deps: GrantDependencies & { clientRepository: ClientRepository }): GrantHandler => {
 	const { config, clientRepository, keyStore } = deps;
